refactor: 헤더 뒤로가기 버튼 클라이언트 컴포넌트로 분리
